# src/app/Models/Siswa.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Siswa extends Model
{
    use HasFactory;

    protected $fillable = [
        'nama',
        'nis',
        'kelas',
        'jurusan',
        'alamat',
    ];

    // Tambahkan relasi ini
    public function absensis()
    {
        return $this->hasMany(Absensi::class);
    }
}